[IMP] Allow resource management on gantt view
- Add a pager

- Add 4 attributes in the gantt view declaration :
1) consolidation :
Specify a field to consolidate. The value of the field will be added to
the parent level :
2) consolidation_max
Specify a maximum value for some group_by. If the maximum is reached,
the consolidation will be red
3) consolidation_exclude :
A field representing a boolean. If the boolean is true, the
consolidation will exclude the record
4) baseline :
Allow to specify some field (start and stop time) to display a baseline

- Read the first level with group_by. Can be used to display unused
  resource

- Create on the fly by clicking on a date

- Add two option attributes :
1) fold_last_level :
Allow to hide the last level in the hierarchy
2) round_dnd_dates :
round the dates while drag and drop.

- css to less

- remove the colors attribute for the gantt declaration